requirements.txt
python-telegram-bot==13.15
requests==2.28.2
PyDictionary==2.0.1
nltk==3.8.1
textblob==0.17.1
python-dotenv==1.0.0